Rename input_port_descriptor.{h,cc}->input_port.{h,cc}.
