+++ systems/controllers/state_feedback_controller_interface.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "drake/common/drake_copyable.h"
-#include "drake/systems/framework/input_port_descriptor.h"
+#include "drake/systems/framework/input_port.h"
 #include "drake/systems/framework/output_port.h"
 
 namespace drake {
